RESUME-ATS/requirements.txt
streamlit
pdfminer.six
python-docx
fpdf
scikit-learn
pandas
matplotlib
